Define SectorObject class for contents to inherit.
This will, by default, create an object with an id, random name, and coordinates.

It provides the to_dict() and generate_name() functions as a base.
