+++ game/objects/planet.py
@@ -1,29 +1,10 @@
 import uuid
-from base import GameObject
 from coordinates import Coordinates
 from random import choice
+from sector import SectorObject
 
-class Planet(GameObject):
+class Planet(SectorObject):
     yaml_tag = "!Planet"
-
-    def __init__(
-        self,
-        name = None,
-        id = str(uuid.uuid4()),
-        coordinates = Coordinates(0,0,0),
-    ):
-        self.name = name if name else self.generate_name()
-        self.id = id
-        self.coordinates = coordinates
-
-        # Parent init should be called at end of __init__
-        super(Planet,self).__init__()
-
-    def to_dict(self):
-        """Override to_dict to handle coordinates."""
-        result = super(Planet,self).to_dict()
-        result['coordinates'] = self.coordinates.to_dict()
-        return result
 
     def generate_name(self):
         """
